+++ wiki/core/markdown/mdx/codehilite.py
@@ -1,9 +1,72 @@
+from __future__ import unicode_literals
+
 import logging
+import re
 
 from markdown.extensions.codehilite import CodeHilite, CodeHiliteExtension
+from markdown.preprocessors import Preprocessor
 from markdown.treeprocessors import Treeprocessor
 
 logger = logging.getLogger(__name__)
+
+
+def highlight(code, config, tab_length, lang=None):
+    code = CodeHilite(
+        code,
+        linenums=config['linenums'],
+        guess_lang=config['guess_lang'],
+        css_class=config['css_class'],
+        style=config['pygments_style'],
+        noclasses=config['noclasses'],
+        tab_length=tab_length,
+        use_pygments=config['use_pygments'],
+        lang=lang,
+    )
+    html = code.hilite()
+    html = """<div class="codehilite-wrap">{}</div>""".format(html)
+    return html
+
+
+class WikiFencedBlockPreprocessor(Preprocessor):
+    """
+    This is a replacement of markdown.extensions.fenced_code which will
+    directly and without configuration options invoke the vanilla CodeHilite
+    extension.
+    """
+    FENCED_BLOCK_RE = re.compile(r'''
+(?P<fence>^(?:~{3,}|`{3,}))[ ]*         # Opening ``` or ~~~
+(\{?\.?(?P<lang>[a-zA-Z0-9_+-]*))?[ ]*  # Optional {, and lang
+# Optional highlight lines, single- or double-quote-delimited
+(hl_lines=(?P<quot>"|')(?P<hl_lines>.*?)(?P=quot))?[ ]*
+}?[ ]*\n                                # Optional closing }
+(?P<code>.*?)(?<=\n)
+(?P=fence)[ ]*$''', re.MULTILINE | re.DOTALL | re.VERBOSE)
+    CODE_WRAP = '<pre>%s</pre>'
+
+    def __init__(self, md):
+        super(WikiFencedBlockPreprocessor, self).__init__(md)
+
+        self.checked_for_codehilite = False
+        self.codehilite_conf = {}
+
+    def run(self, lines):
+        """ Match and store Fenced Code Blocks in the HtmlStash. """
+
+        text = "\n".join(lines)
+        while 1:
+            m = self.FENCED_BLOCK_RE.search(text)
+            if m:
+                lang = ''
+                if m.group('lang'):
+                    lang = m.group('lang')
+                html = highlight(m.group('code'), self.config, self.markdown.tab_length, lang=lang)
+                placeholder = self.markdown.htmlStash.store(html, safe=True)
+                text = '%s\n%s\n%s' % (text[:m.start()],
+                                       placeholder,
+                                       text[m.end():])
+            else:
+                break
+        return text.split("\n")
 
 
 class HiliteTreeprocessor(Treeprocessor):
@@ -14,18 +77,7 @@ class HiliteTreeprocessor(Treeprocessor):
         blocks = root.iter('pre')
         for block in blocks:
             if len(block) == 1 and block[0].tag == 'code':
-                code = CodeHilite(
-                    block[0].text,
-                    linenums=self.config['linenums'],
-                    guess_lang=self.config['guess_lang'],
-                    css_class=self.config['css_class'],
-                    style=self.config['pygments_style'],
-                    noclasses=self.config['noclasses'],
-                    tab_length=self.markdown.tab_length,
-                    use_pygments=self.config['use_pygments']
-                )
-                html = code.hilite()
-                html = """<div class="codehilite-wrap">{}</div>""".format(html)
+                html = highlight(block[0].text, self.config, self.markdown.tab_length)
                 placeholder = self.markdown.htmlStash.store(html, safe=True)
                 # Clear codeblock in etree instance
                 block.clear()
@@ -53,5 +105,17 @@ class WikiCodeHiliteExtension(CodeHiliteExtension):
             )
             del md.treeprocessors["hilite"]
         md.treeprocessors.add("hilite", hiliter, "<inline")
+
+        if "fenced_code_block" in md.preprocessors:
+            logger.warning(
+                "Replacing existing 'fenced_code_block' extension - please remove "
+                "'fenced_code_block' or 'extras' from WIKI_MARKDOWN_KWARGS"
+            )
+            del md.preprocessors["fenced_code_block"]
+        hiliter = WikiFencedBlockPreprocessor(md)
+        hiliter.config = self.getConfigs()
+        md.preprocessors.add('fenced_code_block',
+                             hiliter,
+                             ">normalize_whitespace")
 
         md.registerExtension(self)