+++ wiki/tests/test_markdown.py
@@ -1,10 +1,7 @@
 from __future__ import absolute_import, unicode_literals
 
-import markdown
-from django.test import TestCase
 from mock import patch
 from wiki.core.markdown import ArticleMarkdown
-from wiki.core.markdown.mdx.responsivetable import ResponsiveTableExtension
 from wiki.models import URLPath
 from wiki.tests.base import ArticleTestBase
 
@@ -29,18 +26,3 @@ class ArticleMarkdownTests(ArticleTestBase):
         )
 
         self.assertEqual(urlpath.article.render(), "<p>&lt;/html&gt;only_this</p>")
-
-
-class ResponsiveTableTests(TestCase):
-
-    def setUp(self):
-        self.md = markdown.Markdown(extensions=[
-            'extra',
-            ResponsiveTableExtension()
-        ])
-        self.md_without = markdown.Markdown(extensions=['extra'])
-
-    def test_wrapping(self):
-        text = '|th|th|\n|--|--|\n|td|td|'
-        expected = '<div class="table-responsive">\n' + self.md_without.convert(text) + '\n</div>'
-        self.assertEqual(self.md.convert(text), expected)